Implement dropShadow option in MenuWidgetText class
A dropShadow field and related methods were added to the MenuWidgetText class. The dropShadow option is now integrated into the constructor to allow for more flexibility when creating text components. The default value for dropShadow is set to false to abstain from giving unexpected results with previous implementations of this class.